backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
dj-database-url
Pillow
requests
jdatetime

daphne
black
django-debug-toolbar
